% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tAgg-package.R, R/tAgg.R
\docType{package}
\name{tAgg}
\alias{tAgg}
\title{tAgg (time Aggregation): Time aggregation of vector, matrices or data.frames.}
\usage{
tAgg(
  time,
  x,
  by = "days",
  keep_sum = FALSE,
  na.action = "propagate",
  na.th = 0.05,
  method = "linear",
  verbose = FALSE
)
}
\arguments{
\item{time}{Time vector in POSIXct.}

\item{x}{A vector, matrix or data.frame with dimensions consistent with
argument \code{time} containing the values to be aggregated.}

\item{by}{Character string specifying the target time interval of the
aggregation.}

\item{keep_sum}{Logical (default is \code{FALSE}). If \code{TRUE}, the time
aggregated values are not divided by the duration of the target time steps.}

\item{na.action}{Character string specifying how missing values are to be handled.
Three options are possible: (1) "propagate" (default) where any missing value is
propagated to its target time step, (2) "ignore" where missing values are treated
as 0s and (3) "threshold" where the proportion specified in argument \code{na.th}
is used as threshold of the maximum proportion (time wise) of accepted missing 
values for a given target time step; if below the threshold the "ignore" rule is 
followed, if above, the "propagate" rule is followed.}

\item{na.th}{Numeric value between 0 and 1 specifying the maximum proportion of
accepted missing value when using the "threshold" rule with argument \code{na.action}.}

\item{method}{Character string specifying how aggregation should be done. Two
options are possible: (1) "linear" (default) where source values are linearly
interpolated and the trapezoidal rule is followed in the aggregation;
(2) "constant" where source values are considered constant.}

\item{verbose}{Logical. Should function progress information be printed in the
console.}
}
\value{
A data.frame with a first column \code{time} and one additional column
if \code{x} is a vector or \code{m} additional columns if \code{x} was matrix or
data.frame with \code{m} columns.
}
\description{
This package contains a single function \code{tAgg} to perform time aggregation
from a source time series to a time aggregated time series. Given a time vector
and a vector/matrix/data.frame of values, it computes the time aggregated
values at a given target time interval (e.g. "days", "month", "years", etc.).

Given a time vector (in POSIXct format) and a vector/matrix/data.frame of
values, this function computes the time aggregated values at a given target
time interval (e.g. "days").
}
\details{
Argument \code{time} should be provided in POSIXct format. If another format
is used, the function will attempt to convert to POSIXct with a warning using
\code{as.POSIXct(time, tz="UTC")}.

The dimensions of \code{time} and \code{x} must be consistent. If argument
\code {time} is of length \code{n}, argument \code{x} can be one of the
following: (1) a vector of length \code{n} or (2) a matrix or data.frame with
\code{n} rows.
}
\author{
Ivan Horner
}
